Book fixes #2 [skip-ci]
Move counter-app to book/examples

Update relative paths in commands and book instructions

Add README to pony examples

pony/alphabet
pony/celsius
