+++ book/examples/pony/alphabet/.gitignore
@@ -1,3 +1,4 @@
 alphabet
 sender
 sent.txt
+alphabet.out
