+++ book/examples/pony/celsius/.gitignore
@@ -1,1 +1,4 @@
 celsius
+celsius.msg
+celsius.out
+sent.txt
