+++ book/examples/pony/celsius/data_gen/.gitignore
@@ -1,0 +1,2 @@
+data_gen
+celsius.msg
