--- book/examples/pony/celsius/generator/.gitignore
@@ -1,1 +1,0 @@
-generator
