--- examples/counter-app/Makefile.example
@@ -1,38 +1,0 @@
-CC=clang++
-PONYC=ponyc
-
-SDIR=cpp
-IDIR=hpp
-ODIR=build
-
-CFLAGS=-Wall -std=c++11 -stdlib=libc++ -I$(IDIR)
-
-_DEPS=Counter.hpp
-DEPS=$(patsubst %,$(IDIR)/%,$(_DEPS))
-
-_OBJ=Counter.o
-OBJ=$(patsubst %,$(ODIR)/%,$(_OBJ))
-
-LIBCOUNTER=$(ODIR)/libcounter.a
-COUNTERAPP=$(ODIR)/counter-app
-
-AR=ar
-ARFLAGS=rs
-
-all: $(COUNTERAPP)
-
-lib: $(LIBCOUNTER)
-
-$(ODIR)/%.o: $(SDIR)/%.cpp $(DEPS)
-	$(CC) --debug -c -o $@ $< $(CFLAGS)
-
-$(LIBCOUNTER): $(OBJ)
-	$(AR) $(ARFLAGS) $@ $^
-
-$(COUNTERAPP): $(LIBCOUNTER) counter-app/main.pony
-	$(PONYC) --debug --export --output=$(ODIR) --path=../../lib:../../lib/wallaroo/cpp_api/cpp/cppapi/build/build/lib:./build counter-app
-
-.PHONY: clean
-
-clean:
-	rm -rf $(ODIR)/*.o $(ODIR)/*.a $(ODIR)/counter-app*
